split graph component
